refactor: 더이상 사용하지 않는 Enum Type 제거 및 테스트 코드 수정
