--- src/main/java/chess/domain/Type.java
@@ -1,26 +1,0 @@
-package chess.domain;
-
-public enum Type {
-    ROOK("R", 5),
-    KNIGHT("N", 2.5),
-    BISHOP("B", 3),
-    QUEEN("Q", 9),
-    KING("K", 0),
-    PAWN("P", 1);
-
-    private final String name;
-    private final double score;
-
-    Type(String name, double score) {
-        this.name = name;
-        this.score = score;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public double getScore() {
-        return score;
-    }
-}
