+++ src/main/java/chess/domain/piece/Queen.java
@@ -37,7 +37,7 @@ public class Queen extends Piece {
     }
 
     private Queen(Color color, String name, double score) {
-        super(color, null, name, score);
+        super(color, name, score);
     }
 
     @Override
