+++ src/main/java/chess/domain/piece/Rook.java
@@ -37,7 +37,7 @@ public class Rook extends Piece {
     }
 
     private Rook(Color color, String name, double score) {
-        super(color, null, name, score);
+        super(color, name, score);
     }
 
     @Override
